Move all shift-related methods to shiftable module
