+++ lib/shiftable.rb
@@ -4,6 +4,7 @@ module Shiftable
   SHIFT = Shift.new
   KEY = SHIFT.key
   DATE = SHIFT.date
+  CHARACTERS = ("a".."z").to_a << " "
 
   def generate_keys(key_string)
     {
@@ -30,5 +31,29 @@ module Shiftable
     keys[:c] += offsets[:c]
     keys[:d] += offsets[:d]
     keys
+  end
+
+  def shift_character(character, shift_letter, key_hash, shifted_string)
+    if CHARACTERS.include?(character.downcase)
+      new_arr = CHARACTERS.rotate(CHARACTERS.index(character.downcase))
+      num = key_hash[shift_letter]
+      num -= 27 until num <= 27
+      num = 0 if num == 27
+      shifted_string << new_arr[num]
+    else
+      shifted_string << character
+    end
+  end
+
+  def unshift_character(character, shift_letter, key_hash, shifted_string)
+    if CHARACTERS.include?(character)
+      new_arr = CHARACTERS.rotate(CHARACTERS.index(character))
+      num = key_hash[shift_letter]
+      num -= 27 until num <= 27
+      location = new_arr[0 - num]
+      shifted_string << location
+    else
+      shifted_string << character
+    end
   end
 end
